Fix: switch current device before initial condition
This solves the bug, manifesting on AMD like:

Memory access fault by GPU node-2. Reason: Page
not present or supervisor privilege.

For future reference, this error is just what
it sounds like. You're trying to read from
memory that's not on the current device. It
should highlight the importance of making sure
to only execute GPU code over buffers allocated
on the current device.