Compile Patch::to_device and into_device in non-GPU mode
These functions will panic if called without the
GPU feature enabled, but the compile guard has to
go somewhere, so it might as well be in this
function rather than in calling code. Calling code
can use an ordinary if-statement to determine if
GPU support is available and this function would
succeed.