Add Setup::initial_primitive_dg provided method
This function enables setting the initial data
for DG solver. The grid patch contains nq * ni
f64's, where nq is the number of primitive
fields, and ni is the number of interior
quadrature points. This representation will be
used by the initial conditions generator, but
the plan is to convert to the cons/weights
representation for the solver, and also as the
format for checkpoint data.