Add wavespeed calculation to the srhd solvers
It's not used right now, they still return 1.0
for the max wavespeed to the driver because in
practice the setups currently using these
solvers have ultra-relativistic speeds. The
expanding mesh is factored in by boosting the
primitive state to the comoving frame. This
only reduces the computed wavespeed by a small
margin, because there's relativistic gas at
small comoving radius. What's strange is that
using the expanding mesh still seems to enable
CFL numbers much larger than unity (~1 - 16),
even when the computed wavespeed is in the
comoving frame.