Change relativistic solvers to use conserved as "solution"
The "solution" means the array used by the solver to
represent its state, i.e. on restarts. Until now the
relativistic solvers used the primitive array as the
solution data. It's actually more natural for the
moving mesh to use the (extensive) conserved
variables, but it's also good for post-processing,
since then the primitives as well as the conserved
energies and masses are available without doing
any extra calculations in the plotting scripts.