Fix driver bug: restart from chkpt.final.pk with new cadence
After this commit, if a final checkpoint is written,
then restarting from it will have the expected behavior
even if a new cadence is chosen. However, a final
checkpoint is no longer written by default. Use the
--final-chkpt flag if you need one (a final checkpoint
means one that's force-written, even if it was not
actually time to write one).