Add a new_kernels.py "idea" file
The new kernel module illustrates ways to improve how
Python code interacts with the C/CUDA extension code.
The basic parts are the same, deferring to either cffi or
cupy based on the execution mode, but the implementation
is much cleaner than what’s in sailfish right now. It
encourages a modular as opposed to monolithic approach to
building solvers, so that generic operations things like
PLM, viscous fluxes, point mass source terms) can be shared
between solvers. Currently all of our three solvers for the
binary problem implement those things in monolithic C
sources for each solver. The new way allows you to write
smaller pieces of C/CUDA code inline with a Python wrapper
function. It also allows you to group a bunch of kernels
into a Python class, which makes sense if the kernels
private utility functions.

The GPU part of the new kernel module is not implemented
yet.