Walk back updates-as-generators for now
The idea looks sleek, but it's less clear than
just plain old functions. The only benefit so
far to generators was that the update function
could re-use buffers between iterations. But
the cost to acquire arrays is quite small (<10%
or so on numpy, and even smaller on cupy) when
the arrays are non-initialized (use empty and
empty_like).