Implement fields-first and fields-last data layout
This option might not be offered in general for all solvers,
i.e. many solvers should be permitted to choose one or the
other, but for most-basic solvers it seems like a good idea
for profiling purposes to see how the two approaches compare
and to continue to maintain both modes.