Enable cylindrical r-z; change convention to r-z-f
This convention enables cylindrical coordinates in
the r-z plane. If the need arises for cylindrical
coordinates in the r-f plane, it will probably be
necessary to write another coordinate system with
r-f-z convention. This is because of implementation
details in how for-loops are written, they have
assumed that any squeezed (unit-length) axes are
at the end of the domain.num_zones parameter. This
convention is otherwise not a problem. There is now
a config validation that num_zones only has squeezed
axes at the end.